extract lock
