+++ jobs/cache_maintenance/tests/conftest.py
@@ -3,7 +3,7 @@
 
 from collections.abc import Iterator
 
-from libcommon.queue import _clean_queue_database
+from libcommon.new_queue.utils import _clean_queue_database
 from libcommon.resources import CacheMongoResource, QueueMongoResource
 from libcommon.simple_cache import _clean_cache_database
 from pytest import MonkeyPatch, fixture
