+++ libs/libcommon/src/libcommon/new_queue/__init__.py
@@ -1,2 +1,3 @@
 # SPDX-License-Identifier: Apache-2.0
 # Copyright 2024 The HuggingFace Authors.
+
