+++ libs/libcommon/tests/test_queue.py
@@ -1,13 +1,7 @@
 # SPDX-License-Identifier: Apache-2.0
 # Copyright 2022 The HuggingFace Authors.
 
-import json
-import os
-import random
-import time
 from datetime import datetime, timedelta
-from multiprocessing import Pool
-from pathlib import Path
 from typing import Optional
 from unittest.mock import patch
 
@@ -20,10 +14,8 @@ from libcommon.queue import (
     EmptyQueueError,
     JobDocument,
     JobTotalMetricDocument,
-    Lock,
     Queue,
     WorkerSizeJobsCountDocument,
-    lock,
 )
 from libcommon.resources import QueueMongoResource
 from libcommon.utils import get_datetime
@@ -547,68 +539,6 @@ def test_queue_get_zombies() -> None:
     assert queue.get_zombies(max_seconds_without_heartbeat=9999999) == []
 
 
-def random_sleep() -> None:
-    MAX_SLEEP_MS = 40
-    time.sleep(MAX_SLEEP_MS / 1000 * random.random())
-
-
-def increment(tmp_file: Path) -> None:
-    random_sleep()
-    with open(tmp_file, "r") as f:
-        current = int(f.read() or 0)
-    random_sleep()
-    with open(tmp_file, "w") as f:
-        f.write(str(current + 1))
-    random_sleep()
-
-
-def locked_increment(tmp_file: Path) -> None:
-    sleeps = [0.05, 0.05, 0.05, 1, 1, 1, 1, 1, 1, 5, 5, 5, 5]
-    with lock(key="test_lock", owner=str(os.getpid()), sleeps=sleeps):
-        increment(tmp_file)
-
-
-def test_lock(tmp_path_factory: pytest.TempPathFactory, queue_mongo_resource: QueueMongoResource) -> None:
-    tmp_file = Path(tmp_path_factory.mktemp("test_lock") / "tmp.txt")
-    tmp_file.touch()
-    max_parallel_jobs = 4
-    num_jobs = 42
-
-    with Pool(max_parallel_jobs, initializer=queue_mongo_resource.allocate) as pool:
-        pool.map(locked_increment, [tmp_file] * num_jobs)
-
-    expected = num_jobs
-    with open(tmp_file, "r") as f:
-        assert int(f.read()) == expected
-    Lock.objects(key="test_lock").delete()
-
-
-def git_branch_locked_increment(tmp_file: Path) -> None:
-    sleeps = [0.05, 0.05, 0.05, 1, 1, 1, 1, 1, 1, 5, 5, 5, 5]
-    dataset = "dataset"
-    branch = "refs/convert/parquet"
-    with lock.git_branch(dataset=dataset, branch=branch, owner=str(os.getpid()), sleeps=sleeps):
-        increment(tmp_file)
-
-
-def test_lock_git_branch(tmp_path_factory: pytest.TempPathFactory, queue_mongo_resource: QueueMongoResource) -> None:
-    tmp_file = Path(tmp_path_factory.mktemp("test_lock") / "tmp.txt")
-    tmp_file.touch()
-    max_parallel_jobs = 5
-    num_jobs = 43
-
-    with Pool(max_parallel_jobs, initializer=queue_mongo_resource.allocate) as pool:
-        pool.map(git_branch_locked_increment, [tmp_file] * num_jobs)
-
-    expected = num_jobs
-    with open(tmp_file, "r") as f:
-        assert int(f.read()) == expected
-    assert Lock.objects().count() == 1
-    assert Lock.objects().get().key == json.dumps({"dataset": "dataset", "branch": "refs/convert/parquet"})
-    assert Lock.objects().get().owner is None
-    Lock.objects().delete()
-
-
 def test_delete_dataset_waiting_jobs(queue_mongo_resource: QueueMongoResource) -> None:
     """
     Test that delete_dataset_waiting_jobs deletes all the waiting jobs for a dataset
@@ -689,11 +619,6 @@ def test_delete_dataset_waiting_jobs(queue_mongo_resource: QueueMongoResource) -
     assert JobDocument.objects(dataset=other_dataset).count() == 2
     assert JobDocument.objects(dataset=other_dataset, status=Status.STARTED).count() == 1
     assert JobDocument.objects(dataset=other_dataset, status=Status.WAITING).count() == 1
-
-    assert len(Lock.objects()) == 2
-    assert len(Lock.objects(key=f"{job_type_1},{dataset},{revision}", owner=None)) == 1
-    assert len(Lock.objects(key=f"{job_type_1},{dataset},{revision}", owner__ne=None)) == 0
-    # ^ does not test much, because at that time, the lock should already have been released
 
 
 DATASET_1 = "dataset_1"
