+++ services/webhook/tests/conftest.py
@@ -5,7 +5,7 @@ from collections.abc import Iterator
 
 import pytest
 from libapi.config import UvicornConfig
-from libcommon.queue import _clean_queue_database
+from libcommon.new_queue.utils import _clean_queue_database
 from libcommon.resources import CacheMongoResource, QueueMongoResource
 from libcommon.simple_cache import _clean_cache_database
 from pytest import MonkeyPatch, fixture
