+++ services/worker/src/worker/job_runners/config/parquet_and_info.py
@@ -80,8 +80,8 @@ from libcommon.exceptions import (
     PreviousStepFormatError,
     UnsupportedExternalFilesError,
 )
+from libcommon.new_queue.lock import lock
 from libcommon.parquet_utils import PART_SUFFIX, PARTIAL_PREFIX
-from libcommon.queue import lock
 from libcommon.simple_cache import get_previous_step_or_raise
 from libcommon.utils import HF_HUB_HTTP_ERROR_RETRY_SLEEPS, retry
 from tqdm.contrib.concurrent import thread_map
