+++ services/worker/tests/conftest.py
@@ -4,7 +4,7 @@
 from collections.abc import Iterator
 from pathlib import Path
 
-from libcommon.queue import _clean_queue_database
+from libcommon.new_queue.utils import _clean_queue_database
 from libcommon.resources import CacheMongoResource, QueueMongoResource
 from libcommon.simple_cache import _clean_cache_database
 from libcommon.storage import (
